requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
boto3==1.34.7
botocore==1.34.7
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
cryptography==41.0.5
cycler==0.11.0
defusedxml==0.7.1
Django==4.2.7
django-crispy-forms==2.1
django-rest-auth==0.9.5
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.5
jmespath==1.0.1
kiwisolver==1.4.2
oauthlib==3.2.2
packaging==21.3
Pillow==9.1.1
psycopg==3.1.16
psycopg-binary==3.1.16
psycopg-pool==3.2.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.18
